+++ server.js
@@ -1,7 +1,7 @@
 const request = require("request");
 const http = require("http");
 const express = require("express");
-const { telegram } = require('./js.js')
+const { telegram } = require('./telegram.js')
 const convert = require('xml-js');
 require('dotenv').config();
 
@@ -9,7 +9,10 @@ const app = express();
 const port = process.env.PORT || 3333;
 
 const server = http.createServer(app);
-
+app.post('/webhook',async (req, res) => {
+  res.status(200); 
+  res.end();
+  })
 app.post('/quiz'+process.env.TELEGRAM_TOKEN,async (req, res) => {
 
     GetQuiz((statement, alternatives, correct_index, inlineKeyboard)=>{
@@ -29,13 +32,16 @@ app.post('/tip'+process.env.TELEGRAM_TOKEN,async (req, res) => {
     telegram('sendMessage', {
         chat_id: process.env.CHAT_ID,
         text: Message,
-        reply_markup: JSON.stringify(inlineKeyboard),
+        reply_markup:JSON.stringify(inlineKeyboard),
         parse_mode:'Markdown'
-        },body=>{console.log('ok:'+body.ok)
+        },body=>{console.log(body)
                   res.write(Message) 
                   res.end();})
     })
  })
+
+
+
 function GetQuiz(cb){
   var options = {
     method: 'GET',
@@ -45,12 +51,8 @@ function GetQuiz(cb){
   request(options, function (error, response, body) {
     if (error) throw new Error(error);
     var result1 = convert.xml2js(body, {compact: false, spaces: 4});
-    console.log(result1.elements[0].elements[100+getRandomInt(800)].elements[0].elements[0].text.split('/')[1]) 
-    const data = result1.elements[0].elements[100+getRandomInt(800)].elements[0].elements[0].text.split('/')[1]
+    const data = result1.elements[0].elements[100+getRandomInt(900)].elements[0].elements[0].text.split('/')[1]
   
-
-
-
   var options = {
     method: 'GET',
     url: `https://${process.env.DEZ_API_TOKEN}.amazonaws.com/trivia-paperx-production/lives/${data}/questions.js`
@@ -60,7 +62,7 @@ function GetQuiz(cb){
     if (error) throw new Error(error);
   
    // console.log(body);
-   const { extraQuestion, finalExtraQuestion, questions, videoDurations, feedbackTimes, winnersTime, extraQuestionTime, videoUrls } = eval(body.split('winnersTime, extraQuestionTime').join('winnersTime, extraQuestionTime,videoUrls'));
+   const {  questions,  videoUrls } = eval(body.split('winnersTime, extraQuestionTime').join('winnersTime, extraQuestionTime,videoUrls'));
   const n = getRandomInt(10)
    console.log(questions[n]);
    console.log(videoUrls[n+2]);
@@ -72,7 +74,13 @@ function GetQuiz(cb){
                 text: 'Explicação',
                 url:videoUrls[n+2]
             }
-        ]
+        ],
+        [
+          {
+              text: 'Comentar 💬',
+              url:"https://t.me/ChatDropsENEM"
+          }
+      ]
     ]
   };
 
@@ -95,30 +103,22 @@ request(options, function (error, response, body) {
     var { tipId, exam, subject, content, url} = JSON.parse(body)[getRandomInt(100)]
   console.log({ tipId, exam, subject, content, url});
 
-  const inlineKeyboard = {
-    inline_keyboard: [
-        [
-            {
-                text: '👍',
-                callback_data: 'upvote-'+tipId 
-            },                {
-                text: '👎',
-                callback_data: 'downvote-'+tipId
-            }
-        ],
-        [
-            {
-                text: 'Entenda tudo desse assunto',
-                url:url
-            }
-        ]
-    ]
-  };
   content = content.split('**').join('*')
   content = content.split('_').join('')
   subject = subject.split(' ').join('_')
   exam = exam.split(' ').join('_')
   var Message = `Daily Tip\n#${subject} #${exam}\n\n${content}`
+
+  const inlineKeyboard = {
+    inline_keyboard: [
+        [
+            {
+                text: 'Comentar 💬',
+                url:"https://t.me/ChatDropsENEM"
+            }
+        ]
+    ]
+  };
   
   cb(inlineKeyboard, Message)
 });
